+++ packages/alchemy/services/cloudflare-state-store/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+    "paths": {
+      "alchemy/*": [
+        "../src/*"
+      ]
+    }
+  }
+}
