+++ packages/alchemy/src/State/LocalState.ts
@@ -3,50 +3,9 @@ import * as FileSystem from "effect/FileSystem";
 import * as Layer from "effect/Layer";
 import * as Path from "effect/Path";
 import type { PlatformError } from "effect/PlatformError";
-import * as Redacted from "effect/Redacted";
 import { decodeFqn, encodeFqn } from "../FQN.ts";
-import { encodeState, reviveState } from "./StateEncoding.ts";
 import { State, StateStoreError, type StateService } from "./State.ts";
-
-const REDACTED_MARKER = "__redacted__";
-
-const encodeState = (value: unknown): unknown => {
-  if (value === null || value === undefined) return value;
-  if (Redacted.isRedacted(value)) {
-    return {
-      [REDACTED_MARKER]: encodeState(Redacted.value(value)),
-    };
-  }
-  if (isResource(value)) {
-    return {
-      id: value.LogicalId,
-      type: value.Type,
-      props: encodeState(value.Props),
-      attr: encodeState(value.Attributes),
-    };
-  }
-  if (Array.isArray(value)) return value.map(encodeState);
-  if (typeof value === "object") {
-    const result: Record<string, unknown> = {};
-    for (const [k, v] of Object.entries(value)) {
-      result[k] = encodeState(v);
-    }
-    return result;
-  }
-  return value;
-};
-
-const reviveState = (_key: string, value: unknown): unknown => {
-  if (
-    value !== null &&
-    typeof value === "object" &&
-    !Array.isArray(value) &&
-    REDACTED_MARKER in value
-  ) {
-    return Redacted.make((value as Record<string, unknown>)[REDACTED_MARKER]);
-  }
-  return value;
-};
+import { encodeState, reviveState } from "./StateEncoding.ts";
 
 export const LocalState = Layer.effect(
   State,
