--- services/cloudflare-state-store/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "include": [
-    "alchemy.run.ts",
-    "src/**/*.ts",
-    "src/**/*.d.ts",
-    "test/**/*.ts"
-  ],
-  "compilerOptions": {
-    "noEmit": true,
-    "rootDir": ".",
-    "module": "Preserve",
-    "moduleResolution": "Bundler",
-    "target": "ESNext"
-  },
-  "references": [
-    {
-      "path": "../../packages/alchemy/tsconfig.json"
-    }
-  ]
-}
